fix : feed repository task와 통합
