+++ src/main/java/com/LetMeDoWith/LetMeDoWith/application/feed/dto/RetrieveSuccessDowithTasksResult.java
@@ -1,6 +1,7 @@
 package com.LetMeDoWith.LetMeDoWith.application.feed.dto;
 
-import com.LetMeDoWith.LetMeDoWith.application.feed.repository.dto.SuccessDowithTaskQueryDto;
+import com.LetMeDoWith.LetMeDoWith.domain.task.repository.dto.SuccessDowithTaskQueryDto;
+
 import java.util.List;
 import java.util.Map;
 
@@ -30,5 +31,6 @@ public record RetrieveSuccessDowithTasksResult(Long totalCount, List<SuccessDowi
             String profileImageUrl,
             String successImageUrl,
             boolean isLiked,
-            Long likeCount) {}
+            Long likeCount) {
+    }
 }
