--- src/main/java/com/LetMeDoWith/LetMeDoWith/application/feed/repository/dto/SuccessDowithTaskQueryDto.java
@@ -1,4 +1,0 @@
-package com.LetMeDoWith.LetMeDoWith.application.feed.repository.dto;
-
-public record SuccessDowithTaskQueryDto(
-        Long id, String title, String nickname, String profileImageUrl, String successImageUrl, Boolean isLiked) {}
