+++ src/main/java/com/LetMeDoWith/LetMeDoWith/application/feed/service/FeedDowithTaskService.java
@@ -1,19 +1,21 @@
 package com.LetMeDoWith.LetMeDoWith.application.feed.service;
 
 import com.LetMeDoWith.LetMeDoWith.application.feed.dto.RetrieveSuccessDowithTasksResult;
-import com.LetMeDoWith.LetMeDoWith.application.feed.repository.FeedDowithTaskQueryRepository;
-import com.LetMeDoWith.LetMeDoWith.application.feed.repository.dto.SuccessDowithTaskQueryDto;
-import java.util.Map;
-import java.util.stream.Collectors;
+import com.LetMeDoWith.LetMeDoWith.domain.task.repository.DowithTaskQueryRepository;
+import com.LetMeDoWith.LetMeDoWith.domain.task.repository.dto.SuccessDowithTaskQueryDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+
+import java.util.List;
+import java.util.Map;
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
 public class FeedDowithTaskService {
 
-    private final FeedDowithTaskQueryRepository feedDowithTaskQueryRepository;
+    private final DowithTaskQueryRepository dowithTaskQueryRepository;
 
     /**
      * DowithTask 성공 이미지 피드 조회
@@ -26,10 +28,10 @@ public class FeedDowithTaskService {
         int offset = (int) pageable.getOffset();
         int limit = pageable.getPageSize();
 
-        var totalCount = feedDowithTaskQueryRepository.countSuccessDowithTasks();
-        var successImages = feedDowithTaskQueryRepository.getSuccessDowithTasks(requestMemberId, offset, limit);
+        long totalCount = dowithTaskQueryRepository.countSuccessDowithTasks();
+        List<SuccessDowithTaskQueryDto> successImages = dowithTaskQueryRepository.getSuccessDowithTasks(requestMemberId, offset, limit);
 
-        Map<Long, Long> dowithTaskLikeCountMap = feedDowithTaskQueryRepository.countDowithTaskLikes(
+        Map<Long, Long> dowithTaskLikeCountMap = dowithTaskQueryRepository.countDowithTaskLikes(
                 successImages.stream().map(SuccessDowithTaskQueryDto::id).collect(Collectors.toSet()));
 
         return RetrieveSuccessDowithTasksResult.of(totalCount, successImages, dowithTaskLikeCountMap);
