+++ src/main/java/com/LetMeDoWith/LetMeDoWith/infrastructure/task/query/DowithTaskQueryRespositoryImpl.java
@@ -1,20 +1,25 @@
 package com.LetMeDoWith.LetMeDoWith.infrastructure.task.query;
 
-import com.LetMeDoWith.LetMeDoWith.domain.task.model.QDowithTask;
-import com.LetMeDoWith.LetMeDoWith.domain.task.model.QDowithTaskRoutine;
-import com.LetMeDoWith.LetMeDoWith.domain.task.model.QDowithTaskSuccess;
-import com.LetMeDoWith.LetMeDoWith.domain.task.model.QTaskCategory;
+import com.LetMeDoWith.LetMeDoWith.domain.member.model.QMember;
+import com.LetMeDoWith.LetMeDoWith.domain.task.enums.DowithTaskStatus;
+import com.LetMeDoWith.LetMeDoWith.domain.task.model.*;
 import com.LetMeDoWith.LetMeDoWith.domain.task.repository.DowithTaskQueryRepository;
 import com.LetMeDoWith.LetMeDoWith.domain.task.repository.dto.DowithTaskDetailQueryDto;
 import com.LetMeDoWith.LetMeDoWith.domain.task.repository.dto.DowithTaskQueryDto;
+import com.LetMeDoWith.LetMeDoWith.domain.task.repository.dto.SuccessDowithTaskQueryDto;
 import com.querydsl.core.types.Projections;
 import com.querydsl.core.types.dsl.Expressions;
+import com.querydsl.jpa.JPAExpressions;
 import com.querydsl.jpa.impl.JPAQueryFactory;
-import java.time.LocalDate;
-import java.util.List;
-import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
+
+import java.time.LocalDate;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import java.util.Set;
+import java.util.stream.Collectors;
 
 @Repository
 @RequiredArgsConstructor
@@ -26,6 +31,8 @@ public class DowithTaskQueryRespositoryImpl implements DowithTaskQueryRepository
     private final QDowithTaskRoutine dowithTaskRoutine = QDowithTaskRoutine.dowithTaskRoutine;
     private final QDowithTaskSuccess dowithTaskSuccess = QDowithTaskSuccess.dowithTaskSuccess;
     private final QTaskCategory taskCategory = QTaskCategory.taskCategory;
+    private final QDowithTaskLike dowithTaskLike = QDowithTaskLike.dowithTaskLike;
+    private final QMember member = QMember.member;
 
     @Override
     public List<DowithTaskQueryDto> getDowithTasks(String memberId, LocalDate startDate, LocalDate endDate) {
@@ -46,7 +53,7 @@ public class DowithTaskQueryRespositoryImpl implements DowithTaskQueryRepository
                         // FeedBack
                         // 개발시
                         // 추가
-                        ))
+                ))
                 .from(dowithTask)
                 .leftJoin(taskCategory)
                 .on(dowithTask.taskCategoryId.eq(taskCategory.id))
@@ -82,7 +89,7 @@ public class DowithTaskQueryRespositoryImpl implements DowithTaskQueryRepository
                         // FeedBack
                         // 개발시
                         // 추가
-                        ))
+                ))
                 .from(dowithTask)
                 .leftJoin(taskCategory)
                 .on(dowithTask.taskCategoryId.eq(taskCategory.id))
@@ -94,5 +101,56 @@ public class DowithTaskQueryRespositoryImpl implements DowithTaskQueryRepository
                 .fetchOne();
 
         return Optional.ofNullable(result);
+    }
+
+    @Override
+    public Long countSuccessDowithTasks() {
+        return queryFactory
+                .select(dowithTask.count())
+                .from(dowithTask)
+                .where(dowithTask.status.eq(DowithTaskStatus.SUCCESS))
+                .fetchOne();
+    }
+
+    @Override
+    public List<SuccessDowithTaskQueryDto> getSuccessDowithTasks(String requestMemberId, int offset, int limit) {
+        return queryFactory
+                .select(Projections.constructor(
+                        SuccessDowithTaskQueryDto.class,
+                        dowithTask.id,
+                        dowithTask.title,
+                        member.nickname,
+                        member.profileImageUrl,
+                        dowithTaskSuccess.imageUrl,
+                        JPAExpressions.selectOne()
+                                .from(dowithTaskLike)
+                                .where(
+                                        dowithTaskLike.dowithTask.eq(dowithTask),
+                                        dowithTaskLike.memberId.eq(requestMemberId))
+                                .exists()))
+                .from(dowithTask)
+                .leftJoin(member)
+                .on(dowithTask.memberId.eq(member.id))
+                .join(dowithTaskSuccess)
+                .on(dowithTaskSuccess.dowithTask.eq(dowithTask))
+                .where(dowithTask.status.eq(DowithTaskStatus.SUCCESS))
+                .orderBy(dowithTaskSuccess.createdAt.desc())
+                .offset(offset)
+                .limit(limit)
+                .fetch();
+    }
+
+    @Override
+    public Map<Long, Long> countDowithTaskLikes(Set<Long> dowithTaskIds) {
+        return queryFactory
+                .select(dowithTask.id, dowithTaskLike.count())
+                .from(dowithTaskLike)
+                .where(dowithTaskLike.dowithTask.id.in(dowithTaskIds))
+                .groupBy(dowithTaskLike.dowithTask.id)
+                .fetch()
+                .stream()
+                .collect(Collectors.toMap(tuple -> tuple.get(dowithTask.id), tuple -> Optional.ofNullable(
+                                tuple.get(dowithTaskLike.count()))
+                        .orElse(0L)));
     }
 }
